Add python script to generate BUILD.gn used by chromium Android project (#43)
Chromium projects use GN as Makefile to build the whole project. Add BUILD.gn for fluent icon, so that we can integrate it to Edge Android project conveniently. There is a GN target named "fluent_icon_resources" in the BUILD.gn. If other module need to use the fluent icon resources, we only need add a dependence to "fluent_icon_resources" in other BUILD.gn file.

Co-authored-by: huilliu <huilliu@micrsofot.com>