Add dotnet-blazor package (#436)
The goal of this 'package' is to be able to create the FluentUI System Icons as optimized plain svg assets in a folder structure (assets name with spaces removed). The result is included as static assets in the official Microsoft (Fast) FluentUI Blazer component library. See the FluentUI Blazore package for reference.

No actual package is created when running the npm build command.