Add a new package to publish a react-native-svg version of this icon package. This is a modified version of the react-icons package designed to output a package compatible with react-native-svg.
To accomplish this we we need to use a later version of SVGR which has some changes with file templates and possibly sizing output that are incompatible with the current react-icons code.

Ideally we might want to produce some common code between the two but I don't want to risk breaking anything in react-icons with this additional package.